refactor(preload): extract shared invokePierCommand envelope helper
5 preload modules (index/git-api/worktree-api/plugin-settings-api/
terminal-status-bar-api) each hand-rolled a byte-identical copy of the
PierCommandResult envelope unwrap. Verified all 5 were verbatim
identical before collapsing them into src/preload/ipc-envelope.ts.
Pure refactor, no behavior or signature change.
